test: measure code coverage for test files
Code coverage is usually seen as the parts of your application code that
are covered by tests. A higher percentage is *one* *indicator* of a
safer application, as more code has been run during the tests. It is
*not* a perfect indicator nor a number that *must* be at 100%, as those
tests could be testing the wrong thing, or nothing at all, or mocking
all important parts and just testing an empty shell...

However, adding code coverage to the tests themselves and ensuring it
stays at 100% is interesting. While some parts of the app could be
written "just in case" or be too complex to test, if you have written a
test: you must mean it! Making sure that all code under the "test"
folder has been run is a simple way to avoid mistakes, such as a test
file not named `*_test.rb` that will not be found by Rails test runner
(happened in [c83f0ef9dc4f9bea78a126443701afa9a3d6fa03][1]), a loop that
should test multiple things but looping over an empty set so it doesn't
test anything (e.g. you query models and check an attribute on each,
but the query is wrong)...
[An interesting video by Anthony Sottile on the topic][2].

This commit also moves the SimpleCov configuration from the
`test_helper.rb` to its own `.simplecov` file, which ensures it is
configured only once even in parallel tests or with multiple frameworks.

[1]: https://github.com/rezoleo/lea5/pull/441/files#diff-1cba8981fca6b8913a8746b7ab404d7110f86cf99448b8076bd0ca1e2d79ea6f
[2]: https://www.youtube.com/watch?v=70T6OxKwxm0